Add files via upload
This Snake Game Mini Project in C is a basic console program with no graphics. You may play the famous "Snake Game" in this project exactly as you would anywhere else. To move the snake, use the up, down, right, and left arrows.

Food is placed at various co-ordinates on the screen for the snake to consume. The snake's length and score will both rise by one element each time it consumes the food.